+++ src/main/java/examples/helloboard/domain/Board.java
@@ -11,7 +11,6 @@ public class Board {
     private Integer idx;
     private String title;
     private String content;
-    private Integer fileIdx;
     private Integer userIdx;
     private String date;
     private Integer topicIdx;
